MAJ de l'organisation structurelle Séparation des deux class principale dans deux fichier différents. Ajout visuel pour permettre le suivit des propriétaire de chaque case en vue de la gestion des groupe de couleur.
Il reste a gérer les groupe de couleurs, les événements, et il faut également trouver un moyen de rendre ca plus joli.
